Update linters and fix new reported issues
On local system, following steps are needed:
```
make install
make init_lint
```
