+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 exclude: ^sync_tests/
 repos:
 - repo: https://github.com/asottile/reorder_python_imports
-  rev: v2.7.1
+  rev: v3.0.1
   hooks:
   - id: reorder-python-imports
     entry: env SETUPTOOLS_USE_DISTUTILS=stdlib reorder-python-imports
@@ -37,7 +37,7 @@ repos:
     - flake8-simplify
     - flake8-use-pathlib
 - repo: https://github.com/asottile/pyupgrade
-  rev: v2.31.0
+  rev: v2.31.1
   hooks:
   - id: pyupgrade
     language_version: python3
